Wave 28: 1967 -> 2109 (+142), bytes 21.03% -> 22.14%
Largest wave the project has had. Three agents, address-locality blocks, six
blocks assigned plus three taken as named extension work. 149 of 150 verified
by exit code; 142 promoted into 71 units; both builds reproduce the ROM.

THE FLOOR WAS THE CORPUS, AGAIN -- AND THE COST PREDICTION WAS BACKWARDS.
Wave 27 recorded that `--block-min-matched 20` reports `blocks: 0` and that
this is a statement about the threshold, not the corpus. Confirmed: at 12 the
same screen returned 28 blocks / 284 candidates / 12,968 bytes. But the
prediction attached to it -- "expect them to cost more per function than wave
27's did, fewer promoted siblings means less vocabulary already written down"
-- is REFUTED. The brief told agents to expect that and to say so if it was
not true. Wave 27 landed 95 at density >=20; wave 28 landed 142 at density
>=12, at a HIGHER per-function hit rate (149/150 vs 95/~110). Blocks that are
a third promoted are not materially harder than blocks that are two thirds
promoted. The ~284 candidates the lowered floor exposes are ordinary work, not
a marginal tail. Do not re-derive this prediction; it has been tested.

HARNESS: `overlap_screen.py --top-blocks` DEFAULTS TO 6 AND TRUNCATES
SILENTLY AT THE BLOCK LEVEL. The header printed `blocks: 28, candidate
functions: 284` above six block listings totalling 120 candidates. 164
candidates were invisible with no `(+N more)` notice -- the row-level
truncation inside a block prints one, the block-level one does not. This
caps a wave's available work at whatever six blocks happen to sort highest,
and it is the same failure class wave 26 caught by comparing two numbers by
hand. Pass `--top-blocks 12` or higher.

INVESTIGATION (W28-B, folded into a matching batch at the 3-agent cap): are
the five parked 97-100% near-misses unit-level artifacts of the per-function
oracle rather than source failures? ANSWER: NO, none of the five -- but the
question was worth asking because two cannot be tested at all, for two
different reasons, and answering it overturned two standing diagnoses.

  - sub_0802C604 and sub_08079EA4: `--unit` EXITS 2 without comparing.
    sub_08079EA4's unit holds five functions and only it has a draft, so its
    99.2% is not evidence about the source at all. ACTIONABLE and bounded:
    draft sub_0807974C, sub_08079B04, sub_08079FAC, sub_0807A0C4, re-run
    `--unit`. Do not spend another attempt on sub_08079EA4's own C first.
  - NEW TOOLING GAP: a unit containing an ASM-RESIDENT member can never be
    drafted whole -- `--unit` demands a draft for every member and has no
    notion of "this member is asm-resident, emit its assembly". That is a
    second cheap fix beside the splitter boundary already recorded for
    sub_0802C604. Until then neither oracle can pass it.
  - The other three run and still fail, and NOT ONE residual sits at a `bl`,
    which rules out the static-helper axis structurally rather than by
    exhaustion.

TWO PARKED DIAGNOSES OVERTURNED. Both had stood as "allocation tie-break, no
source-level lever", which is a conclusion that stops future work:
  - sub_0801A6C0 is not allocation noise. It is constant materialisation:
    the ROM uses ONE zero for both `str [r1,#4]` and `strh [r1,#0]`; the
    draft materialises a second. The source writes both fields from one zero.
    Closest of the five to closable and it names its own fix.
  - sub_080815C0's twelve bytes are one register PAIR swapped -- `(5-i)*2`
    and `i-4` take r6/r5 in the ROM and r5/r6 in the draft. Equal-priority
    pseudos allocate in RTL-creation order, so the lever is the order the two
    expressions are first REFERENCED in the source.

STRUCT-TAG COLLISION RECURRED (the wave-20 mode) and promote.py caught it:
three drafts merging into c_08043304.c each named `struct Unk43304`, two with
{unk00, filler_04, unk06} and one with {unk00} alone. All three byte-match
ALONE, so trymatch is blind to it by construction; merged it is a hard
redefinition. Fixed by giving the truncated draft the refined body -- all
three re-verified. Expect this whenever promote.py reports several functions
into one file.

New parks, both with ruled-out axes recorded: sub_0803550C (29.6%, wrong
shape not a near miss, survived a from-scratch re-derivation) and
sub_08043630 (95.5%, one register coalesce, return type and pointer-local
spellings both probed). sub_0806E7FC was parked at 96.2% mid-wave and then
CLOSED on a follow-up round -- the agent removed its own parked entry, which
is the correct sequence.

+++ src/decomp/c_08017860.c
@@ -7,14 +7,7 @@
  * sub_08017860 @ 0x08017860
  */
 
-/* Promoted from assembly; each function below is byte-for-byte
- * identical to the original. Order is address order and must
- * stay that way -- the linker places this file's .text as one
- * contiguous block at 0x08017860.
- * sub_08017860 @ 0x08017860
- */
-
-u8 sub_08017860(int index)
+int sub_08017860(int index)
 {
     return gUnknown_0200C420.unk20[index];
 }
